output implemented

+++ determinant.java
@@ -1,12 +1,12 @@
 public class determinant {
-    public int Determinant(int[][] matrix) {
-        int det = 0; //determinant
+    public double Determinant(double[][] matrix) {
+        double det = 0; //determinant
         int side = matrix.length; //side length of matrix
         if(side==1) {
             det = matrix[0][0];
         } else { // side greater than 1
             for(int i = 0; i < side; i++) {
-                int detTerm = matrix[0][i] * Determinant(Submatrix(matrix,0,i));
+                double detTerm = matrix[0][i] * Determinant(Submatrix(matrix,0,i));
                 if(i%2==0) {
                     det += detTerm;
                 }else{
@@ -16,10 +16,10 @@ public class determinant {
         }
         return det;
     }
-    public double[][] Inverse(int[][]matrix)
+    public double[][] Inverse(double[][]matrix)
     {
-        int det = Determinant(matrix);
-        int[][] cofactor = new int[matrix.length][matrix.length];
+        double det = Determinant(matrix);
+        double[][] cofactor = new double[matrix.length][matrix.length];
         for(int i = 0; i < matrix.length; i++)//cofactor matrix
         {
             for(int k = 0; k < matrix.length; k++)
@@ -38,10 +38,10 @@ public class determinant {
         }
         return inverse;
     }
-    public int[][] Submatrix(int[][]matrix, int ignoredRow, int ignoredColumn) 
+    public double[][] Submatrix(double[][]matrix, int ignoredRow, int ignoredColumn) 
     {
         int subLength = matrix.length - 1; //length of submatrix
-        int[][] submatrix = new int[subLength][subLength];
+        double[][] submatrix = new double[subLength][subLength];
         for(int row = 0; row < matrix.length; row++)
         {
             if(row == ignoredRow) continue;
